05_django_cbv/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
